move elastic fixtures out of conftest
