+++ tests/test_elastic.py
@@ -1,6 +1,8 @@
 import pytest
 import torch
 
+from typing import Any
+from ase.build import bulk, crystal
 from torch_sim.elastic import (
     BravaisType,
     calculate_elastic_moduli,
@@ -10,6 +12,7 @@ from torch_sim.neighbors import vesin_nl_ts
 from torch_sim.state import SimState
 from torch_sim.unbatched.unbatched_optimizers import frechet_cell_fire
 from torch_sim.units import UnitConversion
+from torch_sim.io import atoms_to_state
 
 
 try:
@@ -18,6 +21,111 @@ try:
     from torch_sim.unbatched.models.mace import UnbatchedMaceModel
 except ImportError:
     pytest.skip("MACE not installed", allow_module_level=True)
+
+
+
+
+
+@pytest.fixture
+def mg_atoms() -> Any:
+    """Create crystalline magnesium using ASE."""
+    return bulk("Mg", "hcp", a=3.17, c=5.14)
+
+
+@pytest.fixture
+def sb_atoms() -> Any:
+    """Create crystalline antimony using ASE."""
+    return bulk("Sb", "rhombohedral", a=4.58, alpha=60)
+
+
+@pytest.fixture
+def tio2_atoms() -> Any:
+    """Create crystalline TiO2 using ASE."""
+    a, c = 4.60, 2.96
+    symbols = ["Ti", "O", "O"]
+    basis = [
+        (0.5, 0.5, 0),  # Ti
+        (0.695679, 0.695679, 0.5),  # O
+    ]
+    return crystal(
+        symbols,
+        basis=basis,
+        spacegroup=136,  # P4_2/mnm
+        cellpar=[a, a, c, 90, 90, 90],
+    )
+
+
+@pytest.fixture
+def ga_atoms() -> Any:
+    """Create crystalline Ga using ASE."""
+    a, b, c = 4.43, 7.60, 4.56
+    symbols = ["Ga"]
+    basis = [
+        (0, 0.344304, 0.415401),  # Ga
+    ]
+    return crystal(
+        symbols,
+        basis=basis,
+        spacegroup=64,  # Cmce
+        cellpar=[a, b, c, 90, 90, 90],
+    )
+
+
+@pytest.fixture
+def niti_atoms() -> Any:
+    """Create crystalline NiTi using ASE."""
+    a, b, c = 2.89, 3.97, 4.83
+    alpha, beta, gamma = 90.00, 105.23, 90.00
+    symbols = ["Ni", "Ti"]
+    basis = [
+        (0.369548, 0.25, 0.217074),  # Ni
+        (0.076622, 0.25, 0.671102),  # Ti
+    ]
+    return crystal(
+        symbols,
+        basis=basis,
+        spacegroup=11,
+        cellpar=[a, b, c, alpha, beta, gamma],
+    )
+
+
+
+@pytest.fixture
+def sb_sim_state(sb_atoms: Any, device: torch.device) -> Any:
+    """Create a basic state from sb_atoms."""
+    return atoms_to_state(sb_atoms, device, torch.float64)
+
+
+
+@pytest.fixture
+def cu_sim_state(cu_atoms: Any, device: torch.device) -> Any:
+    """Create a basic state from cu_atoms."""
+    return atoms_to_state(cu_atoms, device, torch.float64)
+
+
+@pytest.fixture
+def mg_sim_state(mg_atoms: Any, device: torch.device) -> Any:
+    """Create a basic state from mg_atoms."""
+    return atoms_to_state(mg_atoms, device, torch.float64)
+
+
+@pytest.fixture
+def tio2_sim_state(tio2_atoms: Any, device: torch.device) -> Any:
+    """Create a basic state from tio2_atoms."""
+    return atoms_to_state(tio2_atoms, device, torch.float64)
+
+
+@pytest.fixture
+def ga_sim_state(ga_atoms: Any, device: torch.device) -> Any:
+    """Create a basic state from ga_atoms."""
+    return atoms_to_state(ga_atoms, device, torch.float64)
+
+
+@pytest.fixture
+def niti_sim_state(niti_atoms: Any, device: torch.device) -> Any:
+    """Create a basic state from niti_atoms."""
+    return atoms_to_state(niti_atoms, device, torch.float64)
+
 
 
 @pytest.fixture
